Refactor utilities into a separate model utils file.
