+++ scripts/model_comparator.py
@@ -8,6 +8,7 @@ import scipy.stats
 import torch
 
 import lm_eval.evaluator
+import lm_eval.models.utils
 from lm_eval import tasks, utils
 
 
@@ -113,7 +114,7 @@ if __name__ == "__main__":
         batch_size=args.batch,
     )
     memory_stats()
-    utils.clear_torch_cache()
+    lm_eval.models.utils.clear_torch_cache()
     eval_logger.info("Memory stats cleared")
     memory_stats()
     results_hf = lm_eval.evaluator.simple_evaluate(
